% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaChecks.R
\name{betasOK}
\alias{betasOK}
\title{betasOK}
\usage{
betasOK(betas)
}
\arguments{
\item{betas}{a potential betas object to check}
}
\value{
boolean
}
\description{
performs checks on Beta ojects to see if the are of the expected form:
a data.frame or matrix with cg probes as rownames and samples names as
column names with numeric values between 0 and 1 for the body
Checks for common errors such as having probe names in the first row
instead of as rownames
}
