% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HannumCoefficients}
\alias{HannumCoefficients}
\title{HannumCoefficients}
\format{An object of class \code{data.frame} with 72 rows and 2 columns.}
\usage{
HannumCoefficients
}
\description{
HannumCoefficients
}
\keyword{datasets}
